Refactor entities.
Update game.sql to use bigint (long) instead of int.

Change dto records to classes.

Change entity id type to Long.

+++ server/src/main/java/com/example/gameserver/controller/GameController.java
@@ -30,7 +30,7 @@ public class GameController {
     }
 
     @GetMapping("/{id}")
-    public GameDto get(@PathVariable Integer id) {
+    public GameDto get(@PathVariable Long id) {
         return gameService.getGame(id);
     }
 
@@ -48,7 +48,7 @@ public class GameController {
 
     @PreAuthorize(Authority.ADMIN_SCOPE)
     @DeleteMapping("/{id}")
-    public void delete(@PathVariable Integer id) {
+    public void delete(@PathVariable Long id) {
         gameService.deleteGame(id);
     }
 }
