+++ server/src/main/java/com/example/gameserver/model/domain/Category.java
@@ -18,10 +18,11 @@ import java.util.Set;
 @Entity
 @Table(name = "category")
 public class Category {
+
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "id", nullable = false)
-    private Integer id;
+    private Long id;
 
     @Size(max = 255)
     @NotNull
