+++ server/src/main/java/com/example/gameserver/repository/DeveloperRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface DeveloperRepository extends JpaRepository<Developer, Integer> {
+public interface DeveloperRepository extends JpaRepository<Developer, Long> {
 }
